creating ibc and light-client sub-crates
Ibc crate includes a handful of structs redefinitions for the cosmos ibc
wasm light client.
They serve as convenient structs to cope with the Go json
serialization/deserialization of `[]byte]` type.
Extracting functions to get LightClientState and LightConsensusState are
included

light-client crate is the wasm contract that runs the IBC LightClient
for Celo on the Cosmos network.
WARNING: it still is a work in progress as Proof verification are not
yet implemented